requirements.txt
opencv-python
numpy
matplotlib
pandas
transformers
torch
Pillow
jiwer
openpyxl